store: hide disabled resources in terminal UI
also moves HUD view rendering into the HUD package

Fixes https://github.com/tilt-dev/tilt/issues/5543
